% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DIV.R
\name{calc_DIV}
\alias{calc_DIV}
\title{Calculate daily infection values}
\usage{
calc_DIV(date_time, Tm, RH, rain, dat)
}
\arguments{
\item{date_time}{POSIX_ct, date time the weather recording was taken}

\item{Tm}{numeric, temperature, in Celsius' at time increment in
\code{date_time}}

\item{RH}{numeric, relative humidity (\%) at time increment in \code{date_time}}

\item{rain}{numeric, volume of rain in millimetres recorded between time recordings}

\item{dat}{data.frame, containing column names "times","temp","rh","rain" with
each of the respective arguments for input. provided as a convenience}
}
\value{
data.table, with probability of infection for each day, between 0 and 1
Undertaken with two methods by Wolf \insertCite{wolf_factors_2005}{cercospoRa}
}
\description{
This function calculates the daily infection values for \emph{Cercospora beticola}
on sugar beet. Functions were adapted from \insertCite{wolf_factors_2005}{cercospoRa}
and \insertCite{wolf_zum_2001}{cercospoRa}
}
\examples{
date_t <- Sys.time() + seq(0, 179 * 60 * 10, (60 * 10))
Tm <- rnorm(180,20,10)
RH <- runif(180,min = 40,90)
rain <- rbinom(180,1,0.1) * runif(180,0.1,20)

DIV1 <- calc_DIV(
date_time = date_t,
Tm = Tm,
RH = RH,
rain = rain
)
}
\references{
\insertAllCited{}
}
