% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_epidemic_onset.R
\name{calc_epidemic_onset}
\alias{calc_epidemic_onset}
\title{Calculate epidemic onset}
\usage{
calc_epidemic_onset(
  c_closure = as.POSIXct("2023-06-01"),
  weather,
  cultivar_sus = 5
)
}
\arguments{
\item{c_closure}{POSIXct formated date to start the model running the model
This is usually at canopy closure (racca)}

\item{weather}{data.table, formated with \code{epiphytoolR::format_weather}}

\item{cultivar_sus}{character, susceptibiliy of the cultivar in "R" resistant,
"S" susceptible, "MR" moderately resistant ect.}
}
\value{
numeric, proportion an epidemic indicating the progress to
}
\description{
Calculate epidemic onset
}
