% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inf_rate.R
\name{calc_inf_rate}
\alias{calc_inf_rate}
\title{Calculate infection rate}
\usage{
calc_inf_rate(Tm, vpd, RH = NA)
}
\arguments{
\item{Tm}{numeric, Temperature}

\item{vpd}{numeric, Vapour pressure deficit}

\item{RH}{numeric, (optional) Relative humidity, when vpd is not available function
will calculate internally}
}
\value{
numeric, probability of infection between 0 and 1
}
\description{
Calculate infection rate
}
\details{
Calculate the infection rate of \emph{Cercospora berticola} in sugar beet leaves.
This function is an attempt to copy the infection rate described by Racca and
Jorg (2007).
}
\examples{
calc_inf_rate(25,0.2)

temp <- seq(5,45, by = 1)
VPD <- seq(0,0.3, by = 0.01)

i_rate <-
  outer(temp,VPD,function(xi,ji){
    calc_inf_rate(Tm = xi, vpd = ji)
     })
persp(temp,VPD,i_rate, theta = 160, ticktype = "detailed")

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\\insertRef\{racca_cercbet_2007\}\{cercospoRa\}
}\if{html}{\out{</div>}}
}
