% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pseudo_maturity.R
\name{calc_pseudo_maturity}
\alias{calc_pseudo_maturity}
\title{Calculate pseudomata maturity}
\usage{
calc_pseudo_maturity(weather, lower_Tm_thres = 10, rain_thresh = 2)
}
\arguments{
\item{weather}{A \code{epiphy.weather} object (an extension of \CRANpkg{data.table})
with weather data formated with the \code{epiphytoolR::format_weather()} function}

\item{lower_Tm_thres}{numeric, lower temperature threshold, which pseduomata don't
mature}

\item{rain_thresh}{numeric, minimum rainfall required to initiate pseudomata
maturity}
}
\value{
numeric, number of days which met the requirement for pseudomata maturity
}
\description{
Calculate pseudomata maturity
}
\examples{
scaddan <-
   system.file("extdata", "scaddan_weather.csv",package = "epiphytoolR")
weather_dat <- read.csv(scaddan)
weather_dat$Local.Time <-
   as.POSIXct(weather_dat$Local.Time, format = "\%Y-\%m-\%d \%H:\%M:\%S",
              tz = "UTC")

weather <- epiphytoolR::format_weather(
   w = weather_dat,
   POSIXct_time = "Local.Time",
   ws = "meanWindSpeeds",
   wd_sd = "stdDevWindDirections",
   rain = "Rainfall",
   temp = "Temperature",
   wd = "meanWindDirections",
   lon = "Station.Longitude",
   lat = "Station.Latitude",
   station = "StationID",
   time_zone = "UTC"
)

calc_pseudo_maturity(weather)

}
