% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sb_emergence.R
\name{calc_sb_emergence}
\alias{calc_sb_emergence}
\title{Estimate sugarbeat emergence}
\usage{
calc_sb_emergence(
  Tm,
  sowing_depth = 5,
  bulk_density = 1.28,
  seed_density = 2.65,
  soil_water_content = 0.7,
  soil_particle_diameter = 0.01,
  mu1 = -4.8,
  mu2 = -8.1,
  alfa = 0.89,
  bta = 0.56,
  Tm_base = 4,
  Tm_crit = 38,
  na.rm = FALSE
)
}
\arguments{
\item{Tm}{numeric, average hourly temperature between sowing and emergence in Celsius.
Or a vector of hourly of temperature means can also be provided.}

\item{sowing_depth}{numeric, given in centimetres. Also considered to be the
internode length. Default = 5}

\item{bulk_density}{numeric, in grams per cm-3. Soil bulk density is defined by
the particles size and pore size.}

\item{seed_density}{numeric, in grams per cm-3. Default defined in publication}

\item{soil_water_content}{numeric, usually between 0.55 (drier) and 0.7 (saturated)}

\item{soil_particle_diameter}{numeric, diameter of soil particle size. This is
not described and guessed. No units were provided either.}

\item{mu1}{model parameter, This is defined by soil properties, such as, soil
moisture, soil particle size and sowing depth. By default we use soil type 3
described in the paper \insertCite{rimaz_predicting_2020}{cercosporaR},
Daneshkadeh soil Seri.}

\item{mu2}{model parameter, See \code{mu1} description}

\item{alfa}{model parameter, See \code{mu1} description}

\item{bta}{model parameter, See \code{mu1} description}

\item{Tm_base}{numeric, lowest temperature which contributes to the development
of the seedling and emergence. Default 6$^o$C}

\item{Tm_crit}{numeric, critical (Highest) temperature which contributes to
the development of the seedling and emergence. Ideal development temperature
is described as between 24 and 28 degrees \insertCite{rimaz_predicting_2020}{cercosporaR}}

\item{na.rm}{logical, remove any NA values when averaging days to emergence.}
}
\value{
numeric, the estimated days to emergence in days. always returns a single
number
}
\description{
This function estimates the time required for sugarbeat emergence as described
by \insertCite{rimaz_predicting_2020}{cercosporaR}
}
\examples{
calc_sb_emergence(27)
calc_sb_emergence(20:40)
}
\references{
\insertAllCited{}
}
