% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spore_rate.R
\name{calc_spore_rate}
\alias{calc_spore_rate}
\title{Calculate sporulation rate}
\usage{
calc_spore_rate(Tm, RH)
}
\arguments{
\item{Tm}{numeric, Temperature}

\item{RH}{numeric, Relative humidity, when vpd is not available function
will calculate internally}
}
\value{
numeric probability of infection between 0 and 1
}
\description{
Calculate sporulation rate
}
\details{
Calculate the sporulation rate of \emph{Cercospora berticola}.
This function is an attempt to copy the infection rate described by Racca and
Jorg (2007).
}
\examples{
calc_spore_rate(25,0.2)

temp <- seq(5,35, by = 1)
RH <- seq(90,100, by = 0.3)

s_rate <-
  outer(temp,RH,function(xi,ji){
    calc_spore_rate(Tm = xi, RH = ji)
     })
persp(temp,RH,s_rate, theta = 315, phi = 20, ticktype = "detailed")

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\\insertRef\{@racca_cercbet_2007\}\{cercosporaR\}
}\if{html}{\out{</div>}}
}
