% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sugarbeet_gdd.R
\name{calc_sugarbeet_gdd}
\alias{calc_sugarbeet_gdd}
\title{Calculate Sugarbeet Growing degree-days}
\usage{
calc_sugarbeet_gdd(
  max_tm,
  min_tm,
  lower_th = 1.1,
  upper_th = 30,
  tm_optim = 26,
  hourly = FALSE
)
}
\arguments{
\item{max_tm}{numeric, maximum daily temperature}

\item{min_tm}{numeric, minimum daily temperature}

\item{lower_th}{numeric, lowest temperature which sugar beet grow at (Celcius)}

\item{upper_th}{numeric, highest temperature which sugar beet grow at (Celcius)}

\item{tm_optim}{numeric, optimum temperature for fastest growth}

\item{hourly}{logical, if \code{TRUE} the function assumes the max_tm and min_tm are
hourly maximums or minimums}
}
\value{
Sum of the growing degree days
}
\description{
Calculate Sugarbeet Growing degree-days
}
\examples{
daily_min <- c(5,7,9,1,-5,-2,0)
daily_max <- c(25,37,29,11,15,12,10)
calc_sugarbeet_gdd(daily_min, daily_max)
}
