% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cercospora_dk.R
\name{cercospora_dk}
\alias{cercospora_dk}
\title{Dispersal kernal for Cercospora berticola}
\usage{
cercospora_dk(x, y, sigma, lambda)
}
\arguments{
\item{x}{lateral coordinates of the location to evaluate the dispersal kernel}

\item{y}{longitude of coordinates of the location to evaluate the dispersal kernel}

\item{sigma}{the standard deviation of the Gaussian component}

\item{lambda}{parameter controlling the exponential decay component}
}
\value{
numeric estimating the distance of the spore travel
}
\description{
Dispersal kernal for Cercospora berticola
}
