% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_period.R
\name{latent_period}
\alias{latent_period}
\title{Calculate Cercospora latent period}
\usage{
latent_period(Tm)
}
\arguments{
\item{Tm}{numeric, Average temperature in degrees Celsius for a hour of the day}
}
\value{
numeric, Latent period for a given hour
}
\description{
Calculate Cercospora latent period
}
\references{
\insertRef{wolf_factors_2005}{cercosporaR}; equation 4
}
