% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_threshold.R
\name{rain_threshold}
\alias{rain_threshold}
\title{Evaluate rainfall threshold}
\usage{
rain_threshold(w, hours = 48, rain_mm = 2)
}
\arguments{
\item{w}{data.table}

\item{hours}{integer, time period in days for which the rainfall threshold is
calculated}

\item{rain_mm}{numeric, total rainfall needed in time period (\code{days}) for rainfall
threshold to be met.}
}
\value{
A \code{epiphy.weather} object (an extension of \CRANpkg{data.table}) with
an additional logical column \code{rain_threshold}.  Read more at \code{?(format_weather)}
The added \code{rain_threshold} indicates whether the rainfall threshold was met.
}
\description{
Evaluate rainfall threshold
}
\examples{
scaddan <-
   system.file("extdata", "scaddan_weather.csv",package = "epiphytoolR")
weather_dat <- read.csv(scaddan)
weather_dat$Local.Time <-
   as.POSIXct(weather_dat$Local.Time, format = "\%Y-\%m-\%d \%H:\%M:\%S",
              tz = "UTC")

weather <-
 epiphytoolR::format_weather(
   w = weather_dat,
   POSIXct_time = "Local.Time",
   ws = "meanWindSpeeds",
   wd_sd = "stdDevWindDirections",
   rain = "Rainfall",
   temp = "Temperature",
   wd = "meanWindDirections",
   lon = "Station.Longitude",
   lat = "Station.Latitude",
   station = "StationID",
   time_zone = "UTC"
)

#weather_out <- rain_threshold(weather)
#weather_out
}
