% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_period.R
\name{temperature_index}
\alias{temperature_index}
\title{Get temperature index}
\usage{
temperature_index(Tm, opt_Tm = 23)
}
\arguments{
\item{Tm}{numeric, temperature for any given hour}

\item{opt_Tm}{numeric, the lowest temperature optimum at which all temperatures
above will have the same (fastest) latent period.}
}
\value{
numeric, proportion representing the speed of the latent period in relation
to the temperature optimum
}
\description{
Temperature index is a proportional representation of the latent period.
Temperatures at or above the optimum temperature for the disease cycle will
yield 1 and the lower the temperature the infinitely longer the latent period.
}
\details{
See \insertCite{wolf_factors_2005}{cercosporaR} and \insertCite{wolf_zum_2001}{cercosporaR}
}
\references{
\insertRef{wolf_factors_2005}{cercosporaR} \cr
\insertRef{wolf_zum_2001}{cercosporaR} \cr
}
