% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weathr}
\alias{weathr}
\title{Weather station data}
\source{
Facundo
A data frame with 8,016 rows and 15 columns:
\describe{
\item{Station}{Identifier}
\item{Datum}{Date}
\item{Stunde}{hour}
\item{Minute}{minute}
\item{T-005}{Temperature at 5cm}
\item{T-200}{Temperature at 200cm}
\item{T-020}{Temperature at 20cm}
\item{F-200}{Relative humidity at 200cm}
...
}
}
\description{
A unformatted weather station dataset from Germany
}
\author{
Facundo
}
