% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_period.R
\name{latent_period}
\alias{latent_period}
\title{Calculate Cercospora latent period}
\usage{
latent_period(Tm, ref = "wolf")
}
\arguments{
\item{Tm}{numeric, Average temperature in degrees Celsius for a hour of the day}

\item{ref}{character, method for calculating latent period. Default is \code{"wolf"}
also available \code{"jarroudi"}. See references for where formulas were used}
}
\value{
numeric, Latent period for a given hour. \code{ref = "jarroudi"} returns
latent period in days.
}
\description{
Calculate Cercospora latent period
}
\details{
Calculates the latent period for \emph{Cercospora beticola} infections on
sugar beet
}
\references{
\insertRef{wolf_factors_2005}{cercospoRa}; equation 4
\insertRef{el_jarroudi_weather-based_2021}{cercospoRa}; equation 1
}
