% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weathr}
\alias{weathr}
\title{Weather station data}
\source{
Climavi One + Climavi Soil, Agvolution GmbH, Gottingen, Germany
A data frame with 8,016 rows and 15 columns:
Weather station that accompanied field trial described in cercospora
manuscript
\describe{
\item{Station}{Identifier}
\item{Datum}{Date}
\item{Stunde}{hour}
\item{Minute}{minute}
\item{T-005}{Temperature in degrees Celcius at 5cm}
\item{T-200}{Temperature in degrees Celcius at 200cm}
\item{T-020}{Temperature in degrees Celcius at 20cm}
\item{F-200}{Percentage Relative humidity at 200cm}
\item{GS200}{Global radiation (W/m2) at 200cm}
\item{WR200}{Wind direction degrees at 200cm}
\item{N100}{Precipitation (mm) at 100cm}
\item{WG200}{Wind Speed (m/s) at 200cm}
...
}
}
\description{
A unformatted weather station dataset from Germany
}
\author{
Institute for Sugar Beet Research - IFZ
}
